+++ app/components/imageButton/index.tsx
@@ -6,10 +6,11 @@ interface ImageButtonProps {
   buttonTitle: string;
   imageBase64: string;
   onPress: () => void;
+  lastItem?: boolean;
 }
 
-const ImageButton: React.FC<ImageButtonProps> = ({ buttonTitle, imageBase64, onPress }) => (
-  <TouchableOpacity onPress={onPress} style={styles.componentView}>
+const ImageButton: React.FC<ImageButtonProps> = ({ buttonTitle, imageBase64, onPress, lastItem = false }) => (
+  <TouchableOpacity onPress={onPress} style={{...styles.componentView, marginRight: lastItem ? 20 : 0 }}>
     <View style={styles.imageView}>
       <Image style={styles.image} source={imageBase64 ? {uri: imageBase64} : require('../../../public/not_found.png')} />
     </View>