+++ app/screens/balance/hooks/useBalance.ts
@@ -1,4 +1,4 @@
-import { useContext } from "react";
+import { useContext, useState } from "react";
 import { NativeStackNavigationProp } from "@react-navigation/native-stack";
 
 import useGlobalBalance from "../../../hooks/useBalance";
@@ -6,6 +6,8 @@ import { AppConfigContext } from "../../../context/appConfig.context";
 import RootStackParamList from "../../../models/interfaces/RootScreensParams";
 
 const useBalance = (_navigation: NativeStackNavigationProp<RootStackParamList, "Balance">) => {
+  const [loading, setLoading] = useState<boolean>(false);
+
   const { appConfig } = useContext(AppConfigContext);
   const { listCategory, listBalance, getAllCurrentValues } = useGlobalBalance();
 
@@ -13,6 +15,8 @@ const useBalance = (_navigation: NativeStackNavigationProp<RootStackParamList, "
     appConfig,
     listCategory,
     listBalance,
+    loading, 
+    setLoading,
     getAllCurrentValues,
   }
 }