+++ app/screens/balance/styles.ts
@@ -3,5 +3,8 @@ import { StyleSheet } from "react-native";
 export const styles = StyleSheet.create({
   flatListMenu: {
     marginBottom: 20
+  },
+  loadingView: {
+    marginTop: 50
   }
 });